So i changed the move, and added a new constructor, added comments too
I'm kinda confused on the implementation, I believe we're going to need a 3d array for the grid because we're modeling a 3d aquarium. Ya know? We have to make sure fish can swim past each other without the program breaking. More than one fish will need to be at a single x,y coord at a time, so we need to have a z. Also passing the grid to the prey every time we have to move is going to take an incredible amount of work, we should have the grid move the pieces for us. It visits each of it's entries, calls "move" on them and then decides if it can take action based on that. I would also like to give the grid object the opportunity to move the prey on it's own if it sees the predator is catching up to it. ya know?